@@ -82,63 +82,51 @@ class _DebuggingControlsState extends State<DebuggingControls>
     required bool resuming,
   }) {
     final isSystemIsolate = controller.isSystemIsolate;
-    return RoundedOutlinedBorder(
-      child: Row(
-        children: [
-          DebuggerButton(
-            title: 'Pause',
-            icon: Codicons.debugPause,
-            autofocus: true,
-            roundedLeftBorder: true,
-            // Disable when paused or selected isolate is a system isolate.
-            onPressed: (isPaused || isSystemIsolate)
-                ? null
-                : () => unawaited(controller.pause()),
-          ),
-          LeftBorder(
-            child: DebuggerButton(
-              title: 'Resume',
-              icon: Codicons.debugContinue,
-              roundedRightBorder: true,
-              // Enable while paused + not resuming and selected isolate is not
-              // a system isolate.
-              onPressed: ((isPaused && !resuming) && !isSystemIsolate)
-                  ? () => unawaited(controller.resume())
-                  : null,
-            ),
-          ),
-        ],
-      ),
+    return RoundedButtonGroup(
+      items: [
+        ButtonGroupItemData(
+          tooltip: 'Pause',
+          icon: Codicons.debugPause,
+          autofocus: true,
+          // Disable when paused or selected isolate is a system isolate.
+          onPressed: (isPaused || isSystemIsolate)
+              ? null
+              : () => unawaited(controller.pause()),
+        ),
+        ButtonGroupItemData(
+          tooltip: 'Resume',
+          icon: Codicons.debugContinue,
+          // Enable while paused + not resuming and selected isolate is not
+          // a system isolate.
+          onPressed: ((isPaused && !resuming) && !isSystemIsolate)
+              ? () => unawaited(controller.resume())
+              : null,
+        ),
+      ],
     );
   }
 
   Widget _stepButtons({required bool canStep}) {
-    return RoundedOutlinedBorder(
-      child: Row(
-        children: [
-          DebuggerButton(
-            title: 'Step Over',
-            icon: Codicons.debugStepOver,
-            roundedLeftBorder: true,
-            onPressed: canStep ? () => unawaited(controller.stepOver()) : null,
-          ),
-          LeftBorder(
-            child: DebuggerButton(
-              title: 'Step In',
-              icon: Codicons.debugStepInto,
-              onPressed: canStep ? () => unawaited(controller.stepIn()) : null,
-            ),
-          ),
-          LeftBorder(
-            child: DebuggerButton(
-              title: 'Step Out',
-              icon: Codicons.debugStepOut,
-              roundedRightBorder: true,
-              onPressed: canStep ? () => unawaited(controller.stepOut()) : null,
-            ),
-          ),
-        ],
-      ),
+    return RoundedButtonGroup(
+      items: [
+        ButtonGroupItemData(
+          label: 'Step Over',
+          icon: Codicons.debugStepOver,
+          onPressed: canStep ? () => unawaited(controller.stepOver()) : null,
+        ),
+        ButtonGroupItemData(
+          label: 'Step In',
+          icon: Codicons.debugStepInto,
+          onPressed: canStep ? () => unawaited(controller.stepIn()) : null,
+        ),
+        ButtonGroupItemData(
+          label: 'Step Out',
+          icon: Codicons.debugStepOut,
+          onPressed: canStep ? () => unawaited(controller.stepOut()) : null,
+        ),
+      ],
+      minScreenWidthForTextBeforeScaling:
+          DebuggingControls.minWidthBeforeScaling,
     );
   }
 
@@ -315,50 +303,4 @@ class ExceptionMode {
   final String id;
   final String name;
   final String description;
-}
-
-@visibleForTesting
-class DebuggerButton extends StatelessWidget {
-  const DebuggerButton({
-    super.key,
-    required this.title,
-    required this.icon,
-    required this.onPressed,
-    this.autofocus = false,
-    this.roundedLeftBorder = false,
-    this.roundedRightBorder = false,
-  });
-
-  final String title;
-  final IconData icon;
-  final VoidCallback? onPressed;
-  final bool autofocus;
-  final bool roundedLeftBorder;
-  final bool roundedRightBorder;
-
-  @override
-  Widget build(BuildContext context) {
-    return DevToolsTooltip(
-      message: title,
-      child: OutlinedButton(
-        autofocus: autofocus,
-        style: OutlinedButton.styleFrom(
-          side: BorderSide.none,
-          shape: RoundedRectangleBorder(
-            borderRadius: BorderRadius.horizontal(
-              left: roundedLeftBorder ? defaultRadius : Radius.zero,
-              right: roundedRightBorder ? defaultRadius : Radius.zero,
-            ),
-          ),
-        ),
-        onPressed: onPressed,
-        child: MaterialIconLabel(
-          label: title,
-          iconData: icon,
-          minScreenWidthForTextBeforeScaling:
-              DebuggingControls.minWidthBeforeScaling,
-        ),
-      ),
-    );
-  }
 }
