@@ -2,17 +2,28 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+import 'dart:async';
+
+import 'package:devtools_shared/devtools_deeplink.dart';
 import 'package:flutter/foundation.dart';
 
+import '../../shared/analytics/analytics.dart' as ga;
+import '../../shared/analytics/constants.dart' as gac;
+import '../../shared/config_specific/server/server.dart' as server;
 import 'deep_links_model.dart';
-import 'fake_data.dart';
+
+typedef _DomainAndPath = ({String domain, String path});
 
 class DeepLinksController {
+  DeepLinksController() {
+    selectedVariantIndex.addListener(_handleSelectedVariantIndexChanged);
+  }
+
   bool get showSpitScreen => showSpitScreenNotifier.value;
 
   List<LinkData> get getLinkDatasByPath {
     final linkDatasByPath = <String, LinkData>{};
-    for (var linkData in linkDatasNotifier.value) {
+    for (var linkData in linkDatasNotifier.value!) {
       linkDatasByPath[linkData.path] = linkData;
     }
     return linkDatasByPath.values.toList();
@@ -20,14 +31,67 @@ class DeepLinksController {
 
   List<LinkData> get getLinkDatasByDomain {
     final linkDatasByDomain = <String, LinkData>{};
-    for (var linkData in linkDatasNotifier.value) {
+    for (var linkData in linkDatasNotifier.value!) {
       linkDatasByDomain[linkData.domain] = linkData;
     }
     return linkDatasByDomain.values.toList();
   }
 
+  final Map<int, AppLinkSettings> _androidAppLinks = <int, AppLinkSettings>{};
+
+  late final selectedVariantIndex = ValueNotifier<int>(0);
+  void _handleSelectedVariantIndexChanged() {
+    linkDatasNotifier.value = null;
+    unawaited(_loadAndroidAppLinks());
+  }
+
+  Future<void> _loadAndroidAppLinks() async {
+    if (!_androidAppLinks.containsKey(selectedVariantIndex.value)) {
+      final variant =
+          selectedProject.value!.androidVariants[selectedVariantIndex.value];
+      await ga.timeAsync(
+        gac.deeplink,
+        gac.AnalyzeFlutterProject.loadAppLinks.name,
+        asyncOperation: () async {
+          final result = await server.requestAndroidAppLinkSettings(
+            selectedProject.value!.path,
+            buildVariant: variant,
+          );
+          _androidAppLinks[selectedVariantIndex.value] = result;
+        },
+      );
+    }
+    _updateLinks();
+  }
+
+  List<LinkData> get _allLinkDatas {
+    final appLinks = _androidAppLinks[selectedVariantIndex.value]?.deeplinks;
+    if (appLinks == null) {
+      return const <LinkData>[];
+    }
+    final domainPathToScheme = <_DomainAndPath, Set<String>>{};
+    for (final appLink in appLinks) {
+      final schemes = domainPathToScheme.putIfAbsent(
+        (domain: appLink.host, path: appLink.path),
+        () => <String>{},
+      );
+      schemes.add(appLink.scheme);
+    }
+    return domainPathToScheme.entries
+        .map(
+          (entry) => LinkData(
+            domain: entry.key.domain,
+            path: entry.key.path,
+            os: [PlatformOS.android],
+            scheme: entry.value.toList(),
+          ),
+        )
+        .toList();
+  }
+
+  final selectedProject = ValueNotifier<FlutterProject?>(null);
   final selectedLink = ValueNotifier<LinkData?>(null);
-  final linkDatasNotifier = ValueNotifier<List<LinkData>>(allLinkDatas);
+  final linkDatasNotifier = ValueNotifier<List<LinkData>?>(null);
   final showSpitScreenNotifier = ValueNotifier<bool>(false);
 
   final _searchContentNotifier = ValueNotifier<String>('');
@@ -35,7 +99,7 @@ class DeepLinksController {
   void _updateLinks() {
     final searchContent = _searchContentNotifier.value;
     final List<LinkData> linkDatas = searchContent.isNotEmpty
-        ? allLinkDatas
+        ? _allLinkDatas
             .where(
               (linkData) => linkData.matchesSearchToken(
                 RegExp(
@@ -45,8 +109,7 @@ class DeepLinksController {
               ),
             )
             .toList()
-        : allLinkDatas;
-
+        : _allLinkDatas;
     linkDatasNotifier.value = linkDatas;
   }
 
