@@ -17,7 +17,7 @@ const paths = <String>[
 final allLinkDatas = <LinkData>[
   for (var path in paths)
     LinkData(
-      os: ['Android', 'iOS'],
+      os: [PlatformOS.android, PlatformOS.ios],
       domain: 'm.shopping.com',
       path: path,
       domainError: true,
@@ -25,14 +25,14 @@ final allLinkDatas = <LinkData>[
     ),
   for (var path in paths)
     LinkData(
-      os: ['iOS'],
+      os: [PlatformOS.ios],
       domain: 'm.french.shopping.com',
       path: path,
       pathError: path.contains('shoe'),
     ),
   for (var path in paths)
     LinkData(
-      os: ['Android'],
+      os: [PlatformOS.android],
       domain: 'm.chinese.shopping.com',
       path: path,
       pathError: path.contains('shoe'),
