@@ -33,9 +33,11 @@ class AndroidChartController extends ChartController {
     // have been toggled to be displayed - yet.
     if (traces.isNotEmpty) {
       final chartDataLength = timestampsLength;
-      final dataLength = _memoryController.memoryTimeline.data.length;
+      final dataLength =
+          _memoryController.controllers.memoryTimeline.data.length;
 
-      final dataRange = _memoryController.memoryTimeline.data.getRange(
+      final dataRange =
+          _memoryController.controllers.memoryTimeline.data.getRange(
         chartDataLength,
         dataLength,
       );
@@ -143,7 +145,7 @@ class MemoryAndroidChartState extends State<MemoryAndroidChart>
   /// Controller attached to the chart.
   AndroidChartController get _chartController => widget.chartController;
 
-  MemoryTimeline get _memoryTimeline => controller.memoryTimeline;
+  MemoryTimeline get _memoryTimeline => controller.controllers.memoryTimeline;
 
   @override
   void initState() {
