@@ -178,7 +178,8 @@ class ChartsValues {
 
   void _getEventData(Map<String, Object> results) {
     // Use the detailed extension events data stored in the memoryTimeline.
-    final eventInfo = controller.memoryTimeline.data[index].memoryEventInfo;
+    final eventInfo =
+        controller.controllers.memoryTimeline.data[index].memoryEventInfo;
 
     if (eventInfo.isEmpty) return;
 
@@ -221,7 +222,8 @@ class ChartsValues {
   }
 
   void _getVMData(Map<String, Object> results) {
-    final HeapSample heapSample = controller.memoryTimeline.data[index];
+    final HeapSample heapSample =
+        controller.controllers.memoryTimeline.data[index];
 
     results[rssJsonName] = heapSample.rss;
     results[capacityJsonName] = heapSample.capacity;
@@ -234,7 +236,7 @@ class ChartsValues {
 
   void _getAndroidData(Map<String, Object> results) {
     final AdbMemoryInfo androidData =
-        controller.memoryTimeline.data[index].adbMemoryInfo;
+        controller.controllers.memoryTimeline.data[index].adbMemoryInfo;
 
     results[adbTotalJsonName] = androidData.total;
     results[adbOtherJsonName] = androidData.other;
