@@ -87,7 +87,7 @@ void debugTraceEventCallback(VoidCallback callback) {
 
 const preCompileShadersDocsUrl = 'https://docs.flutter.dev/perf/shader';
 
-const impellerWikiUrl = 'https://github.com/flutter/flutter/wiki/Impeller';
+const impellerDocsUrl = 'https://docs.flutter.dev/perf/impeller';
 
 extension TraceEventExtension on TraceEvent {
   bool get isThreadNameEvent =>
