@@ -3,6 +3,7 @@
 // found in the LICENSE file.
 
 import 'dart:async';
+import 'dart:math' as math;
 
 import 'package:devtools_app_shared/service.dart';
 import 'package:devtools_app_shared/ui.dart';
@@ -59,7 +60,10 @@ class ServiceExtensionButtonGroup extends StatefulWidget {
 }
 
 class _ServiceExtensionButtonGroupState
-    extends State<ServiceExtensionButtonGroup> with AutoDisposeMixin {
+    extends State<ServiceExtensionButtonGroup>
+    with
+        AutoDisposeMixin<ServiceExtensionButtonGroup>,
+        MainIsolateChangeMixin<ServiceExtensionButtonGroup> {
   late List<ExtensionState> _extensionStates;
 
   @override
@@ -70,6 +74,9 @@ class _ServiceExtensionButtonGroupState
     // which would be more natural.
     _initExtensionState();
   }
+
+  @override
+  void _onMainIsolateChanged() => _initExtensionState();
 
   void _initExtensionState() {
     _extensionStates = [for (var e in widget.extensions) ExtensionState(e)];
@@ -82,12 +89,12 @@ class _ServiceExtensionButtonGroupState
       final state = serviceConnection.serviceManager.serviceExtensionManager
           .getServiceExtensionState(extensionName);
       extension.isSelected = state.value.enabled;
-
       addAutoDisposeListener(state, () {
         setState(() {
           extension.isSelected = state.value.enabled;
         });
       });
+
       // Track whether the extension is actually exposed by the VM.
       final listenable = serviceConnection
           .serviceManager.serviceExtensionManager
@@ -418,25 +425,31 @@ class _ServiceExtensionToggle extends ServiceExtensionWidget {
 }
 
 class _ServiceExtensionToggleState extends State<_ServiceExtensionToggle>
-    with ServiceExtensionMixin, AutoDisposeMixin {
+    with
+        ServiceExtensionMixin,
+        AutoDisposeMixin<_ServiceExtensionToggle>,
+        MainIsolateChangeMixin<_ServiceExtensionToggle> {
   bool value = false;
 
   @override
   void initState() {
     super.initState();
+    _initExtensionState();
+  }
+
+  @override
+  void _onMainIsolateChanged() => _initExtensionState();
+
+  void _initExtensionState() {
     final state = serviceConnection.serviceManager.serviceExtensionManager
         .getServiceExtensionState(widget.service.extension);
-
     value = state.value.enabled;
 
-    addAutoDisposeListener(
-      state,
-      () {
-        setState(() {
-          value = state.value.enabled;
-        });
-      },
-    );
+    addAutoDisposeListener(state, () {
+      setState(() {
+        value = state.value.enabled;
+      });
+    });
   }
 
   @override
@@ -511,7 +524,10 @@ class ServiceExtensionCheckbox extends ServiceExtensionWidget {
 }
 
 class _ServiceExtensionCheckboxState extends State<ServiceExtensionCheckbox>
-    with ServiceExtensionMixin, AutoDisposeMixin {
+    with
+        ServiceExtensionMixin,
+        AutoDisposeMixin<ServiceExtensionCheckbox>,
+        MainIsolateChangeMixin<ServiceExtensionCheckbox> {
   /// Whether this checkbox value is set to true.
   ///
   /// This notifier listens to extension state changes from the service manager
@@ -524,7 +540,13 @@ class _ServiceExtensionCheckboxState extends State<ServiceExtensionCheckbox>
   @override
   void initState() {
     super.initState();
+    _initExtensionState();
+  }
 
+  @override
+  void _onMainIsolateChanged() => _initExtensionState();
+
+  void _initExtensionState() {
     if (serviceConnection.serviceManager.serviceExtensionManager
         .isServiceExtensionAvailable(widget.serviceExtension.extension)) {
       final state = serviceConnection.serviceManager.serviceExtensionManager
@@ -665,7 +687,10 @@ class ServiceExtensionCheckboxGroupButton extends StatefulWidget {
 }
 
 class _ServiceExtensionCheckboxGroupButtonState
-    extends State<ServiceExtensionCheckboxGroupButton> with AutoDisposeMixin {
+    extends State<ServiceExtensionCheckboxGroupButton>
+    with
+        AutoDisposeMixin<ServiceExtensionCheckboxGroupButton>,
+        MainIsolateChangeMixin<ServiceExtensionCheckboxGroupButton> {
   static const _hoverYOffset = 10.0;
 
   /// Whether this button should have the enabled state, which makes the
@@ -682,6 +707,13 @@ class _ServiceExtensionCheckboxGroupButtonState
   @override
   void initState() {
     super.initState();
+    _initExtensionState();
+  }
+
+  @override
+  void _onMainIsolateChanged() => _initExtensionState();
+
+  void _initExtensionState() {
     _extensionStates = List.filled(widget.extensions.length, false);
     for (int i = 0; i < widget.extensions.length; i++) {
       final extension = widget.extensions[i];
@@ -698,6 +730,7 @@ class _ServiceExtensionCheckboxGroupButtonState
     _enabled.value = _isEnabled();
 
     if (widget.forceShowOverlayController != null) {
+      cancelStreamSubscriptions();
       autoDisposeStreamSubscription(
         widget.forceShowOverlayController!.stream.listen(
           (_) => _insertOverlay(context),
@@ -747,7 +780,11 @@ class _ServiceExtensionCheckboxGroupButtonState
   /// The overlay will appear directly below the button, and will be dismissed
   /// if there is a click outside of the list of toggles.
   void _insertOverlay(BuildContext context) {
-    final offset = _calculateOverlayPosition(widget.overlayWidth, context);
+    final width = math.min(
+      widget.overlayWidth,
+      MediaQuery.of(context).size.width - 2 * denseSpacing,
+    );
+    final offset = _calculateOverlayPosition(width, context);
     _overlay?.remove();
     Overlay.of(context).insert(
       _overlay = OverlayEntry(
@@ -767,7 +804,7 @@ class _ServiceExtensionCheckboxGroupButtonState
                     child: _ServiceExtensionCheckboxGroupOverlay(
                       description: widget.overlayDescription,
                       extensions: widget.extensions,
-                      width: widget.overlayWidth,
+                      width: width,
                       customExtensionUi: widget.customExtensionUi,
                     ),
                   ),
@@ -785,7 +822,7 @@ class _ServiceExtensionCheckboxGroupButtonState
         Overlay.of(context).context.findRenderObject() as RenderBox;
     final box = context.findRenderObject() as RenderBox;
 
-    final maxX = overlayBox.size.width - width;
+    final maxX = overlayBox.size.width - width - denseSpacing;
     final maxY = overlayBox.size.height;
 
     final offset = box.localToGlobal(
@@ -794,7 +831,7 @@ class _ServiceExtensionCheckboxGroupButtonState
     );
 
     return Offset(
-      offset.dx.clamp(0.0, maxX),
+      offset.dx.clamp(denseSpacing, maxX),
       offset.dy.clamp(0.0, maxY),
     );
   }
@@ -1060,6 +1097,28 @@ class ServiceExtensionIcon extends StatelessWidget {
       color: color,
     );
   }
+}
+
+mixin MainIsolateChangeMixin<T extends StatefulWidget>
+    on State<T>, AutoDisposeMixin<T> {
+  static const _mainIsolateListenerId = 'mainIsolateListener';
+
+  @override
+  void initState() {
+    super.initState();
+    addAutoDisposeListener(
+      serviceConnection.serviceManager.isolateManager.mainIsolate,
+      () {
+        setState(() {
+          cancelListeners(excludeIds: [_mainIsolateListenerId]);
+          _onMainIsolateChanged();
+        });
+      },
+      _mainIsolateListenerId,
+    );
+  }
+
+  void _onMainIsolateChanged();
 }
 
 // TODO(jacobr): port these classes to Flutter.
