@@ -704,25 +704,6 @@ class BlankHeader extends StatelessWidget implements PreferredSizeWidget {
   Size get preferredSize => Size.zero;
 }
 
-/// Button to export data.
-///
-/// * `minScreenWidthForTextBeforeScaling`: The minimum width the button can be before the text is
-///    omitted.
-/// * `onPressed`: The callback to be called upon pressing the button.
-class ExportButton extends GaDevToolsButton {
-  ExportButton({
-    required super.gaScreen,
-    super.key,
-    super.onPressed,
-    super.minScreenWidthForTextBeforeScaling,
-    super.tooltip = 'Export data',
-  }) : super(
-          icon: Icons.file_download,
-          label: 'Export',
-          gaSelection: gac.export,
-        );
-}
-
 /// Button to open related information / documentation.
 ///
 /// [tooltip] specifies the hover text for the button.
@@ -995,10 +976,7 @@ class CenteredMessage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Center(
-      child: Text(
-        message,
-        style: Theme.of(context).textTheme.titleLarge,
-      ),
+      child: Text(message),
     );
   }
 }
@@ -1412,6 +1390,51 @@ class MoreInfoLink extends StatelessWidget {
   }
 }
 
+class LinkIconLabel extends StatelessWidget {
+  const LinkIconLabel({
+    super.key,
+    required this.icon,
+    required this.link,
+    required this.color,
+  });
+
+  final IconData icon;
+  final Link link;
+  final Color? color;
+
+  @override
+  Widget build(BuildContext context) {
+    return InkWell(
+      onTap: _onLinkTap,
+      child: Row(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          Icon(
+            icon,
+            size: defaultIconSize,
+            color: color,
+          ),
+          const SizedBox(width: densePadding),
+          Padding(
+            padding: const EdgeInsets.only(bottom: densePadding),
+            child: RichText(
+              text: TextSpan(
+                text: link.display,
+                style: Theme.of(context).linkTextStyle.copyWith(color: color),
+              ),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+
+  void _onLinkTap() {
+    unawaited(launchUrl(link.url));
+    ga.select(link.gaScreenName, link.gaSelectedItemDescription);
+  }
+}
+
 class LinkTextSpan extends TextSpan {
   LinkTextSpan({
     required Link link,
@@ -1657,34 +1680,8 @@ class CheckboxSetting extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final theme = Theme.of(context);
-
-    Widget textContent = RichText(
-      overflow: TextOverflow.visible,
-      text: TextSpan(
-        text: title,
-        style: enabled ? theme.regularTextStyle : theme.subtleTextStyle,
-      ),
-    );
-
-    if (description != null) {
-      textContent = Row(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          textContent,
-          Expanded(
-            child: RichText(
-              maxLines: 2,
-              overflow: TextOverflow.ellipsis,
-              text: TextSpan(
-                text: ' • $description',
-                style: theme.subtleTextStyle,
-              ),
-            ),
-          ),
-        ],
-      );
-    }
-    final content = Row(
+    Widget checkboxAndTitle = Row(
+      mainAxisSize: MainAxisSize.min,
       children: [
         NotifierCheckbox(
           notifier: notifier,
@@ -1703,17 +1700,57 @@ class CheckboxSetting extends StatelessWidget {
           checkboxKey: checkboxKey,
         ),
         Flexible(
-          child: textContent,
+          child: RichText(
+            overflow: TextOverflow.visible,
+            maxLines: 2,
+            text: TextSpan(
+              text: title,
+              style: enabled ? theme.regularTextStyle : theme.subtleTextStyle,
+            ),
+          ),
         ),
       ],
     );
-    if (tooltip != null && tooltip!.isNotEmpty) {
-      return DevToolsTooltip(
-        message: tooltip,
-        child: content,
-      );
+    if (description == null) {
+      checkboxAndTitle = Expanded(child: checkboxAndTitle);
     }
-    return content;
+    return maybeWrapWithTooltip(
+      tooltip: tooltip,
+      child: Row(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          checkboxAndTitle,
+          if (description != null) ...[
+            Expanded(
+              child: Padding(
+                padding: const EdgeInsets.only(top: denseSpacing),
+                child: Row(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    RichText(
+                      text: TextSpan(
+                        text: ' • ',
+                        style: theme.subtleTextStyle,
+                      ),
+                    ),
+                    Flexible(
+                      child: RichText(
+                        maxLines: 4,
+                        overflow: TextOverflow.ellipsis,
+                        text: TextSpan(
+                          text: description,
+                          style: theme.subtleTextStyle,
+                        ),
+                      ),
+                    ),
+                  ],
+                ),
+              ),
+            ),
+          ],
+        ],
+      ),
+    );
   }
 }
 
