@@ -248,7 +248,7 @@ void main() {
         expect(find.byType(AnalysisView), findsOneWidget);
         expect(
           find.text(
-            'Dart AOT snapshot: lib/src/app_size/stub_data/new_v8.dart - 7/28/2020 1:29 PM',
+            'Dart AOT snapshot: lib/src/app_size/stub_data/new_v8.dart - 7/28/2020 1:29 PM',
           ),
           findsOneWidget,
         );
@@ -359,9 +359,10 @@ void main() {
         expect(find.text('No File Selected'), findsNothing);
 
         expect(find.byType(DiffView), findsOneWidget);
+
         expect(
           find.text(
-            'Diffing Dart AOT snapshots: lib/src/app_size/stub_data/old_v8.dart - 7/28/2020 1:29 PM (OLD)    vs    (NEW) lib/src/app_size/stub_data/new_v8.dart - 7/28/2020 1:29 PM',
+            'Diffing Dart AOT snapshots: lib/src/app_size/stub_data/old_v8.dart - 7/28/2020 1:29 PM (OLD)    vs    (NEW) lib/src/app_size/stub_data/new_v8.dart - 7/28/2020 1:29 PM',
           ),
           findsOneWidget,
         );
