@@ -2,9 +2,9 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+import 'package:codicon/codicon.dart';
 import 'package:devtools_app/devtools_app.dart';
 import 'package:devtools_app/src/screens/debugger/codeview.dart';
-import 'package:devtools_app/src/screens/debugger/controls.dart';
 import 'package:devtools_app/src/screens/debugger/debugger_model.dart';
 import 'package:devtools_app/src/shared/diagnostics/primitives/source_location.dart';
 import 'package:devtools_app_shared/ui.dart';
@@ -15,6 +15,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 import 'package:vm_service/vm_service.dart';
 
+import '../test_infra/utils/debugger_utils.dart';
 import '../test_infra/utils/test_utils.dart';
 
 void main() {
@@ -88,20 +89,20 @@ void main() {
       await tester.pump();
 
       expect(
-        findDebuggerButtonWithTitle('Pause'),
+        findDebuggerButtonWithIcon(Codicons.debugPause),
         findsOneWidget,
       );
-      final pause = getWidgetFromFinder<DebuggerButton>(
-        findDebuggerButtonWithTitle('Pause'),
+      final pause = getWidgetFromFinder<OutlinedButton>(
+        findDebuggerButtonWithIcon(Codicons.debugPause),
       );
       expect(pause.onPressed, isNull);
 
       expect(
-        findDebuggerButtonWithTitle('Resume'),
+        findDebuggerButtonWithIcon(Codicons.debugContinue),
         findsOneWidget,
       );
-      final resume = getWidgetFromFinder<DebuggerButton>(
-        findDebuggerButtonWithTitle('Resume'),
+      final resume = getWidgetFromFinder<OutlinedButton>(
+        findDebuggerButtonWithIcon(Codicons.debugContinue),
       );
       expect(resume.onPressed, isNotNull);
     },
@@ -182,10 +183,6 @@ final _stackFrame2 = StackFrameAndSourcePosition(
   ),
   position: const SourcePosition(line: _stackFrame2Line, column: 1),
 );
-
-Finder findDebuggerButtonWithTitle(String title) => find.byWidgetPredicate(
-      (Widget widget) => widget is DebuggerButton && widget.title == title,
-    );
 
 Finder findStackFrameWithText(String text) => find.byWidgetPredicate(
       (Widget widget) =>
