@@ -32,11 +32,15 @@ void main() {
       await controller.collectRasterStats();
     });
 
-    Future<void> pumpRenderingLayerVisualizer(WidgetTester tester) async {
+    Future<void> pumpRasterStatsView(
+      WidgetTester tester, {
+      bool impellerEnabled = false,
+    }) async {
       await tester.pumpWidget(
         wrap(
           RasterStatsView(
             rasterStatsController: controller,
+            impellerEnabled: impellerEnabled,
           ),
         ),
       );
@@ -45,7 +49,7 @@ void main() {
 
     testWidgets('renders in empty state', (WidgetTester tester) async {
       controller.clearData();
-      await pumpRenderingLayerVisualizer(tester);
+      await pumpRasterStatsView(tester);
 
       expect(find.byType(LayerSnapshotTable), findsNothing);
       expect(find.byType(LayerImage), findsNothing);
@@ -58,7 +62,7 @@ void main() {
     });
 
     testWidgets('renders with data', (tester) async {
-      await pumpRenderingLayerVisualizer(tester);
+      await pumpRasterStatsView(tester);
 
       expect(find.byType(LayerSnapshotTable), findsOneWidget);
       expect(find.richText('Layer'), findsOneWidget);
@@ -79,8 +83,23 @@ void main() {
       );
     });
 
+    testWidgets('renders for Impeller', (WidgetTester tester) async {
+      controller.clearData();
+      await pumpRasterStatsView(tester, impellerEnabled: true);
+
+      expect(find.byType(LayerSnapshotTable), findsNothing);
+      expect(find.byType(LayerImage), findsNothing);
+      expect(
+        find.richTextContaining(
+          'The Raster Stats tool is not currently available for the '
+          'Impeller backend.',
+        ),
+        findsOneWidget,
+      );
+    });
+
     testWidgets('can change layer selection', (tester) async {
-      await pumpRenderingLayerVisualizer(tester);
+      await pumpRasterStatsView(tester);
 
       final rasterStats = controller.rasterStats.value!;
       final layers = rasterStats.layerSnapshots;
