@@ -93,8 +93,8 @@ class MemoryDefaultScene extends Scene {
       profilePaneController: profileController,
     )
       ..offline = true
-      ..memoryTimeline.offlineData.clear()
-      ..memoryTimeline.offlineData.addAll(memoryJson.data);
+      ..controllers.memoryTimeline.offlineData.clear()
+      ..controllers.memoryTimeline.offlineData.addAll(memoryJson.data);
   }
 
   @override
@@ -152,7 +152,6 @@ AdaptedHeapData _createHeap(Map<String, int> classToInstanceCount) {
   return AdaptedHeapData(
     objects,
     rootIndex: rootIndex,
-    isolateId: '',
   );
 }
 
