@@ -203,9 +203,11 @@ Color alternatingColorForIndex(int index, ColorScheme colorScheme) {
 /// override the default DevTools themes with.
 ///
 /// A value of 0.5 would result in all colours being considered light/dark, and
-/// a value of 0.1 allowing around only the 10% darkest/lightest colours by
+/// a value of 0.12 allowing around only the 12% darkest/lightest colours by
 /// Flutter's luminance calculation.
-const _lightDarkLuminanceThreshold = 0.1;
+/// 12% was chosen becaues VS Code's default light background color is #f3f3f3
+/// which is a little under 11%.
+const _lightDarkLuminanceThreshold = 0.12;
 
 bool isValidDarkColor(Color? color) {
   if (color == null) {
