@@ -3,10 +3,10 @@
 // found in the LICENSE file.
 
 import 'dart:async';
-import 'dart:io';
 
 import 'package:args/command_runner.dart';
 import 'package:cli_util/cli_logging.dart';
+import 'package:io/io.dart';
 import 'package:path/path.dart' as path;
 
 import '../model.dart';
@@ -23,7 +23,7 @@ class PubGetCommand extends Command {
         _onlyMainFlag,
         negatable: false,
         help: 'Only execute on the top-level `devtools/packages/devtools_*` '
-            'packages',
+            'packages and any of their subdirectories',
       );
   }
 
@@ -42,7 +42,8 @@ class PubGetCommand extends Command {
     }
 
     final log = Logger.standard();
-    final repo = DevToolsRepo.getInstance()!;
+    final repo = DevToolsRepo.getInstance();
+    final processManager = ProcessManager();
     final packages = repo.getPackages();
 
     final upgrade = argResults![_upgradeFlag];
@@ -55,33 +56,29 @@ class PubGetCommand extends Command {
 
     for (Package p in packages) {
       final packagePathParts = path.split(p.relativePath);
-      final isMainPackage = packagePathParts.length == 2 &&
+      final isMainPackageOrSubdirectory = packagePathParts.length >= 2 &&
           packagePathParts.first == 'packages' &&
           packagePathParts[1].startsWith('devtools_');
-      if (onlyMainPackages && !isMainPackage) continue;
+      if (onlyMainPackages && !isMainPackageOrSubdirectory) continue;
 
       final progress = log.progress('  ${p.relativePath}');
 
-      final process = await Process.start(
-        sdk.flutterToolPath,
-        ['pub', command],
+      final process = await processManager.runProcess(
+        CliCommand.flutter(
+          'pub $command',
+          // Run all so we can see the full set of results instead of stopping
+          // on the first error.
+          throwOnException: false,
+        ),
         workingDirectory: p.packagePath,
       );
-      final stderr = process.stderr;
 
-      final exitCode = await process.exitCode;
-
+      final exitCode = process.exitCode;
       if (exitCode == 0) {
         progress.finish(showTiming: true);
       } else {
         failureCount++;
-
-        // Display stderr when pub get goes wrong.
-        final List<List<int>> err = await stderr.toList();
-        final errorOutput = convertProcessOutputToString(err, '    ');
         progress.finish(message: 'failed (exit code $exitCode)');
-
-        log.stderr(log.ansi.error(errorOutput));
       }
     }
 
