@@ -2,8 +2,13 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+import 'package:args/args.dart';
 import 'package:args/command_runner.dart';
 import 'package:devtools_tool/commands/fix_goldens.dart';
+import 'package:devtools_tool/commands/generate_code.dart';
+import 'package:devtools_tool/commands/sync.dart';
+import 'package:devtools_tool/commands/update_flutter_sdk.dart';
+import 'package:io/io.dart';
 
 import 'commands/analyze.dart';
 import 'commands/list.dart';
@@ -18,13 +23,26 @@ class DevToolsCommandRunner extends CommandRunner {
   DevToolsCommandRunner()
       : super('devtools_tool', 'A repo management tool for DevTools.') {
     addCommand(AnalyzeCommand());
-    addCommand(RepoCheckCommand());
+    addCommand(FixGoldensCommand());
+    addCommand(GenerateCodeCommand());
     addCommand(ListCommand());
     addCommand(PubGetCommand());
-    addCommand(RollbackCommand());
-    addCommand(UpdateDartSdkDepsCommand());
     addCommand(ReleaseHelperCommand());
+    addCommand(RepoCheckCommand());
+    addCommand(RollbackCommand());
+    addCommand(SyncCommand());
+    addCommand(UpdateDartSdkDepsCommand());
     addCommand(UpdateDevToolsVersionCommand());
-    addCommand(FixGoldensCommand());
+    addCommand(UpdateFlutterSdkCommand());
+  }
+
+  @override
+  Future runCommand(ArgResults topLevelResults) async {
+    try {
+      return await super.runCommand(topLevelResults);
+    } finally {
+      // Closes stdin for the entire program.
+      await sharedStdIn.terminate();
+    }
   }
 }
